\name{repeated}
\title{Non-normal Repeated Measurements Library}
\description{
\code{\link{biv.betab}}		Bivariate Beta-binomial Regression

\code{\link{biv.binom}}		Marginal Bivariate Binomial Regression

\code{\link{capture}}			Capture-recapture Model

\code{\link{catmiss}}			Marginal Probabilities for Categorical Repeated Measures with Missing Data

\code{\link{chidden}}			Continuous-Time Hidden Markov Chain Models

\code{\link{gar}}			Generalized Autoregression

\code{\link{glmm}}			Generalized Linear Mixed Models

\code{\link{gnlmm}}			Generalized Nonlinear Mixed Models

\code{\link{hidden}}			Discrete-Time Hidden Markov Chain Models

\code{\link{kalcount}}		Generalized Repeated Measurements Models for Counts

\code{\link{kalseries}}		Generalized Repeated Measurements Models for Continuous Variables

\code{\link{logitord}}		Ordinal Random Effects Models with Dropouts

\code{\link{marg.hom}}		Marginal Homogeneity Model

\code{\link{nbkal}}			Negative Binomial Model with Kalman Update
}
