\name{repeated}
\title{Non-normal Repeated Measurements Library}
\alias{repeated}
\description{
\code{\link[repeated]{binnest}}			Binary Random Effects Model with Two Levels of Nesting

\code{\link[repeated]{biv.betab}}		Bivariate Beta-binomial Regression

\code{\link[repeated]{biv.binom}}		Marginal Bivariate Binomial Regression

\code{\link[repeated]{capture}}			Capture-recapture Model

\code{\link[repeated]{catmiss}}			Marginal Probabilities for Categorical Repeated Measures with Missing Data

\code{\link[repeated]{chidden}}			Continuous-Time Hidden Markov Chain Models

\code{\link[repeated]{cmcre}}		Continuous-time Two-state Markov Process with Random Effect

\code{\link[repeated]{cphidden}}			Changepoint Location using a Continuous-time Two-state Hidden Markov Chain

\code{\link[repeated]{gar}}			Generalized Autoregression

\code{\link[repeated]{gausscop}}	Multivariate Gaussian Copula with Arbitrary Marginals

\code{\link[repeated]{gnlmix}}			Generalized Nonlinear Regression with a Random Parameter

\code{\link[repeated]{glmm}}			Generalized Linear Mixed Models

\code{\link[repeated]{gnlmm}}			Generalized Nonlinear Mixed Models

\code{\link[repeated]{gnlmm3}}			Generalized Nonlinear Mixed Models for Three-parameter Distributions

\code{\link[repeated]{hidden}}			Discrete-Time Hidden Markov Chain Models

\code{\link[repeated]{hnlmix}}			Generalized Nonlinear Regression using h-likelihood for a Random Parameter

\code{\link[repeated]{kalcount}}		Generalized Repeated Measurements Models for Counts

\code{\link[repeated]{kalseries}}		Generalized Repeated Measurements Models for Continuous Variables

\code{\link[repeated]{logitord}}		Ordinal Random Effects Models with Dropouts

\code{\link[repeated]{marg.hom}}		Marginal Homogeneity Model

\code{\link[repeated]{nbkal}}			Negative Binomial Model with Kalman Update
}
\keyword{documentation}
